chore(tests): fix broken tests
